[mstp] Fix for passing internal index to stpsync
Changes to fix wrongly passing mst internal index to stpsync api-s, which
expect instance id. For MSTP, since the internal index and MST instance
ids are not one-to-one mapped (for example, for CIST, the MST instance
id 0 but the internal index 64). With these changes, the index are
converted to MST instance id and passed to stpsync api-s

Signed-off-by: vedganes <veda.ganesan@nokia.com>